miscellaneous fixes i've made
OverhealDecay decays extra health down to your maximum over time.  The higher your health, the faster each point of health decays.  This is necessary because of the Psychic Harmonium and Psychic Sizer, which allow players to get infinite overheal.
FixGoldG11 fixes the HK G11 Gold Edition's ironsight accuracy being 10x as wide as it should be, which resulted in the scope having less accuracy than the hipfire.